refactor(authentication): rename user domain to authentication domain
Move reducer inside hooks folder
Separate reducer/actions from initial/handlers
