+++ src/hooks/useAuthentication/authHandlers.js
@@ -1,0 +1,12 @@
+export const initialState = {
+  username: '',
+  company: '',
+}
+
+export function loadUserData([username, company]) {
+  return {username, company}
+}
+
+export function unloadUserData() {
+  return {...initialState}
+}
